+++ js/util.js
@@ -1,5 +1,3 @@
-import { Author, Price, Guests, Rooms, Location, TYPES, CHECKINS, CHECKOUTS, FEATURES, PHOTOS, SIMILAR_ITEM_COUNT } from './data.js';
-
 const getRandomInt = (min, max) => {
 
   if (min >= 0 && max > min) {
@@ -43,32 +41,9 @@ const getUniqRandomLengthArray = (array) => {
   return array.slice(getRandomInt(0, array.length - 1));
 }
 
-const createItem = () => {
-  const latitude = getRandomFloat(Location.X.MIN, Location.X.MAX);
-  const longitude = getRandomFloat(Location.Y.MIN, Location.Y.MAX);
-
-  return {
-    Author: {
-      avatar: `'img/avatars/user0${getRandomInt(Author.MIN, Author.MAX)}.png'`,
-    },
-    offer: {
-      title: 'Специальное предложение!',
-      address: `${latitude}, ${longitude}`,
-      Price: getRandomInt(Price.MIN, Price.MAX),
-      type: getRandomArrayElement(TYPES),
-      Rooms: getRandomInt(Rooms.MIN, Rooms.MAX),
-      Guests: getRandomInt(Guests.MIN, Guests.MAX),
-      checkin: getRandomArrayElement(CHECKINS),
-      checkout: getRandomArrayElement(CHECKOUTS),
-      features: getUniqRandomLengthArray(FEATURES),
-      description: 'Отличный выбор для семьи из трёх человек',
-      photos: getUniqRandomLengthArray(PHOTOS),
-    },
-    location: {
-      x: latitude,
-      y: longitude,
-    },
-  }
-};
-
-const similarItems = new Array(SIMILAR_ITEM_COUNT).fill(null).map(() => createItem());
+export {
+  getRandomInt,
+  getRandomFloat,
+  getRandomArrayElement,
+  getUniqRandomLengthArray
+}
